CRUD Operation of playlist by Admin

+++ Online_Music_System/src/Componets/ManagePlaylist/Playlists.js
@@ -2,46 +2,43 @@ import React, { useState, useEffect } from 'react'
 import { BrowserRouter as Router, useNavigate, Switch, Route, Link } from 'react-router-dom';
 
 export default function Playlists() {
-    const [song, setSong] = useState([])
-    const [sname, setSname] = useState()
-    const fetchSong = async () => {
-        let response = await fetch(`http://localhost:4099/song/fetchallsongs/${"all"}`,
+    const account = JSON.parse(localStorage.getItem("account"))
+    const [playlists, setPlaylists] = useState([])
+    const nav = useNavigate();
+
+    const fetchPlaylists = async () => {
+        let response = await fetch(`http://localhost:4099/playlist/fetchadminplaylists/${"all"}`,
             {
                 method: 'GET',
-                headers: { 'Content-Type': 'application/json' },
-
+                headers: { 'Content-Type': 'application/json'},
             }
         );
         response = await response.json();
-
         if (response.success) {
-            setSong(response.songs);
+            setPlaylists(response.playlists);
         }
-        return response;
     }
 
-    const handleOnClick = async (id) => {
+    const handleEditClick = (p) => {
+        nav('/admin/createplaylist', { state: { playlist: p }})
+    }
+    const handleDeleteClick = async (id) => {
+        const playlist = playlists.find(p => p._id === id)
+        if(window.confirm(`Are you sure you want to delete "${playlist.playlistName}"?`)) {
 
-        let response = await fetch("http://localhost:4099/song/deletesong",
+            let response = await fetch("http://localhost:4099/playlist/deleteplaylist",
             {
                 method: 'DELETE',
                 headers: { 'Content-Type': 'application/json' },
                 body: JSON.stringify({ id })
             }
-        );
-        response = await response.json();
-        fetchSong();
-        return response;
-
+            );
+            response = await response.json();
+            fetchPlaylists();
+        }
+        
     }
-    const nav = useNavigate();
-
-    const navigate = (id) => {
-        nav('/editsong', { state: { id: id } })
-    }
-
     useEffect(() => {
-        const account = JSON.parse(localStorage.getItem("account"))
         if (!account) {
 
             nav('/signin')
@@ -49,16 +46,74 @@ export default function Playlists() {
         if (account && account.user.isAdmin === false) {
             nav('/')
         }
-        fetchSong();
+        fetchPlaylists();
     }, [])
     return (
         <div>
             <div className="container">
-                <Link to="/createplaylist">
-                    <button className="btn btn-success  my-4 mx-auto">+ Create PlayList</button>
+                <h2 className="my-2">Playlists</h2>
+                <hr />
+                <Link to="/admin/createplaylist">
+                    <button className="btn btn-success  my-2 mx-auto">+ Create PlayList</button>
                 </Link>
-            </div>
-            <div>
+            
+
+                {
+                    playlists.map(p => {
+                        return (
+                            <div className="accordion my-2" key={p._id}>
+                                <div className="accordion-item">
+                                    <h2 className="accordion-header">
+                                        <button className="accordion-button collapsed border border-secondary shadow-none" type="button" data-bs-toggle="collapse" data-bs-target={`#${p.playlistName.split(" ").join("-")}`} aria-expanded="false" aria-controls="flush-collapseOne">
+                                            {p.playlistName}
+                                        </button>
+                                    </h2>
+                                    <div id={p.playlistName.split(" ").join("-")} className="accordion-collapse collapse" aria-labelledby="flush-headingOne" data-bs-parent="#accordionFlushExample">
+                                        <div className="accordion-body">
+                                            <table className="table table-striped">
+                                                <thead>
+                                                    <tr>
+                                                        <th>#</th>
+                                                        <th>Song</th>
+                                                        <th>Artist</th>
+                                                    </tr>
+                                                </thead>
+                                                <tbody>
+                                                    {
+                                                        p.songs.map((s, index) => {
+                                                            return (
+                                                                <tr key={s._id}>
+                                                                    <td>{index + 1}</td>
+                                                                    <td>
+                                                                        {s.songName}
+                                                                        {
+                                                                            (s.movieName != "AlbumSong") ?
+                                                                                <span className="">&nbsp;(From "{s.movieName}")</span> : <span></span>
+
+                                                                        }
+                                                                    </td>
+                                                                    <td>{s.singerName}</td>
+                                                                </tr>
+                                                            )
+                                                        })
+                                                    }
+                                                </tbody>
+                                            </table>
+                                            <div className="my-2">
+                                                <button type="button" className="btn btn-success me-4" onClick={() => {handleEditClick(p)}}>
+                                                    Edit
+                                                </button>
+                                                <button type="button" className="btn btn-danger" onClick={() => {handleDeleteClick(p._id)}}>
+                                                    Delete
+                                                </button>
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>
+                        )
+                    })
+                }
             </div>
         </div>
     )
